feat(whiteboard): extract components to separate files
